chore(rollup): only slightly broken config
Due to the way rollup-common-js works, it does not interoperate nicely
with out mixture between ES imports and require('external') statements.

This would require a fix in rollup-common-js. But even with this fix,
there may be plenty of issues we might need to workaround beyond the
horizon (such as rollup silently not esifying certain files; cf.
didi/lib/injector).